Replace the UnionFS
