+++ lib/virtual-fs/src/fs/mod.rs
@@ -5,7 +5,6 @@ pub mod empty;
 pub mod host;
 #[cfg(feature = "native-fs")]
 pub mod native;
-// pub mod new_union;
 pub mod overlay;
 pub mod passthru;
 pub mod tmp;
